Moved the different sections on home page into their separate components.
