+++ components/home/AboutUsSection.jsx
@@ -1,0 +1,148 @@
+import Button from "@/components/Button";
+import styles from "@/styles/home/Home.module.css";
+
+export default function AboutUsSection() {
+  return (
+    <div className={styles["header-content"]}>
+      <div className={styles["about-container"]}>
+        <div className={styles["about-content"]}>
+          <p className={styles["about-title"]}>About Us</p>
+          <p className={styles["about-subtitle"]}>
+            OUR SOCIETY CONSISTS OF FIVE TEAMS.
+          </p>
+          <p className={styles["about-description"]}>
+            Lorem ipsum odor amet, consectetuer adipiscing elit. Dui volutpat
+            torquent netus est augue. Commodo risus vel adipiscing, est
+            convallis nostra. Class fringilla etiam dis tellus suscipit
+            adipiscing fusce curabitur. Proin pellentesque nascetur consectetur
+            convallis, varius quisque. Fermentum ac libero et maximus primis
+            litora massa orci. Fames auctor maecenas . <br />
+            <br />
+            Sagittis himenaeos magna parturient himenaeos dictumst ligula. Amet
+            maximus per diam tempor, egestas ullamcorper. Etiam penatibus urna
+            nulla praesent volutpat hac. Morbi
+          </p>
+          <Button
+            text={"Meet the teams"}
+            color={"trans-white"}
+            fullWidth={true}
+          />
+        </div>
+        <div className={styles["about-col1"]}>
+          <div className={styles["about-box1"]}>
+            <svg
+              width="39"
+              height="27"
+              viewBox="0 0 39 27"
+              fill="none"
+              xmlns="http://www.w3.org/2000/svg"
+            >
+              <path
+                d="M27.5417 9.16667L32.8026 6.01009C34.4219 5.03848 35.2316 4.55268 35.8978 4.61479C36.4787 4.66894 37.0087 4.969 37.3539 5.43926C37.75 5.97856 37.75 6.92277 37.75 8.81121V17.6888C37.75 19.5771 37.75 20.5214 37.3539 21.0608C37.0087 21.531 36.4787 21.8311 35.8978 21.8852C35.2316 21.9473 34.4219 21.4616 32.8026 20.49L27.5417 17.3333M7.53333 25.5H21.0083C23.2952 25.5 24.4387 25.5 25.3122 25.0549C26.0804 24.6635 26.7052 24.0388 27.0966 23.2705C27.5417 22.3971 27.5417 21.2535 27.5417 18.9667V7.53333C27.5417 5.24646 27.5417 4.10301 27.0966 3.22954C26.7052 2.4612 26.0804 1.83653 25.3122 1.44506C24.4387 1 23.2952 1 21.0083 1H7.53333C5.24646 1 4.10301 1 3.22954 1.44506C2.4612 1.83653 1.83653 2.4612 1.44506 3.22954C1 4.10301 1 5.24644 1 7.53333V18.9667C1 21.2535 1 22.3971 1.44506 23.2705C1.83653 24.0388 2.4612 24.6635 3.22954 25.0549C4.10301 25.5 5.24644 25.5 7.53333 25.5Z"
+                stroke="white"
+                strokeWidth="2"
+                strokeLinecap="round"
+                strokeLinejoin="round"
+              />
+            </svg>
+            <p className={styles["about-team-name"]}>AMV Team</p>
+            <p>
+              Lorem ipsum odor amet, consectetuer adipiscing elit. Odio iaculis
+              primis nisi imperd
+            </p>
+          </div>
+          <div className={styles["about-box1"]}>
+            <svg
+              width="39"
+              height="27"
+              viewBox="0 0 39 27"
+              fill="none"
+              xmlns="http://www.w3.org/2000/svg"
+            >
+              <path
+                d="M27.5417 9.16667L32.8026 6.01009C34.4219 5.03848 35.2316 4.55268 35.8978 4.61479C36.4787 4.66894 37.0087 4.969 37.3539 5.43926C37.75 5.97856 37.75 6.92277 37.75 8.81121V17.6888C37.75 19.5771 37.75 20.5214 37.3539 21.0608C37.0087 21.531 36.4787 21.8311 35.8978 21.8852C35.2316 21.9473 34.4219 21.4616 32.8026 20.49L27.5417 17.3333M7.53333 25.5H21.0083C23.2952 25.5 24.4387 25.5 25.3122 25.0549C26.0804 24.6635 26.7052 24.0388 27.0966 23.2705C27.5417 22.3971 27.5417 21.2535 27.5417 18.9667V7.53333C27.5417 5.24646 27.5417 4.10301 27.0966 3.22954C26.7052 2.4612 26.0804 1.83653 25.3122 1.44506C24.4387 1 23.2952 1 21.0083 1H7.53333C5.24646 1 4.10301 1 3.22954 1.44506C2.4612 1.83653 1.83653 2.4612 1.44506 3.22954C1 4.10301 1 5.24644 1 7.53333V18.9667C1 21.2535 1 22.3971 1.44506 23.2705C1.83653 24.0388 2.4612 24.6635 3.22954 25.0549C4.10301 25.5 5.24644 25.5 7.53333 25.5Z"
+                stroke="white"
+                strokeWidth="2"
+                strokeLinecap="round"
+                strokeLinejoin="round"
+              />
+            </svg>
+            <p className={styles["about-team-name"]}>AMV Team</p>
+            <p>
+              Lorem ipsum odor amet, consectetuer adipiscing elit. Odio iaculis
+              primis nisi imperd
+            </p>
+          </div>
+        </div>
+        <div className={styles["about-col1"]}>
+          <div className={styles["about-box1"]}>
+            <svg
+              width="39"
+              height="27"
+              viewBox="0 0 39 27"
+              fill="none"
+              xmlns="http://www.w3.org/2000/svg"
+            >
+              <path
+                d="M27.5417 9.16667L32.8026 6.01009C34.4219 5.03848 35.2316 4.55268 35.8978 4.61479C36.4787 4.66894 37.0087 4.969 37.3539 5.43926C37.75 5.97856 37.75 6.92277 37.75 8.81121V17.6888C37.75 19.5771 37.75 20.5214 37.3539 21.0608C37.0087 21.531 36.4787 21.8311 35.8978 21.8852C35.2316 21.9473 34.4219 21.4616 32.8026 20.49L27.5417 17.3333M7.53333 25.5H21.0083C23.2952 25.5 24.4387 25.5 25.3122 25.0549C26.0804 24.6635 26.7052 24.0388 27.0966 23.2705C27.5417 22.3971 27.5417 21.2535 27.5417 18.9667V7.53333C27.5417 5.24646 27.5417 4.10301 27.0966 3.22954C26.7052 2.4612 26.0804 1.83653 25.3122 1.44506C24.4387 1 23.2952 1 21.0083 1H7.53333C5.24646 1 4.10301 1 3.22954 1.44506C2.4612 1.83653 1.83653 2.4612 1.44506 3.22954C1 4.10301 1 5.24644 1 7.53333V18.9667C1 21.2535 1 22.3971 1.44506 23.2705C1.83653 24.0388 2.4612 24.6635 3.22954 25.0549C4.10301 25.5 5.24644 25.5 7.53333 25.5Z"
+                stroke="white"
+                strokeWidth="2"
+                strokeLinecap="round"
+                strokeLinejoin="round"
+              />
+            </svg>
+            <p className={styles["about-team-name"]}>AMV Team</p>
+            <p>
+              Lorem ipsum odor amet, consectetuer adipiscing elit. Odio iaculis
+              primis nisi imperd
+            </p>
+          </div>
+          <div className={styles["about-box1"]}>
+            <svg
+              width="39"
+              height="27"
+              viewBox="0 0 39 27"
+              fill="none"
+              xmlns="http://www.w3.org/2000/svg"
+            >
+              <path
+                d="M27.5417 9.16667L32.8026 6.01009C34.4219 5.03848 35.2316 4.55268 35.8978 4.61479C36.4787 4.66894 37.0087 4.969 37.3539 5.43926C37.75 5.97856 37.75 6.92277 37.75 8.81121V17.6888C37.75 19.5771 37.75 20.5214 37.3539 21.0608C37.0087 21.531 36.4787 21.8311 35.8978 21.8852C35.2316 21.9473 34.4219 21.4616 32.8026 20.49L27.5417 17.3333M7.53333 25.5H21.0083C23.2952 25.5 24.4387 25.5 25.3122 25.0549C26.0804 24.6635 26.7052 24.0388 27.0966 23.2705C27.5417 22.3971 27.5417 21.2535 27.5417 18.9667V7.53333C27.5417 5.24646 27.5417 4.10301 27.0966 3.22954C26.7052 2.4612 26.0804 1.83653 25.3122 1.44506C24.4387 1 23.2952 1 21.0083 1H7.53333C5.24646 1 4.10301 1 3.22954 1.44506C2.4612 1.83653 1.83653 2.4612 1.44506 3.22954C1 4.10301 1 5.24644 1 7.53333V18.9667C1 21.2535 1 22.3971 1.44506 23.2705C1.83653 24.0388 2.4612 24.6635 3.22954 25.0549C4.10301 25.5 5.24644 25.5 7.53333 25.5Z"
+                stroke="white"
+                strokeWidth="2"
+                strokeLinecap="round"
+                strokeLinejoin="round"
+              />
+            </svg>
+            <p className={styles["about-team-name"]}>AMV Team</p>
+            <p>
+              Lorem ipsum odor amet, consectetuer adipiscing elit. Odio iaculis
+              primis nisi imperd
+            </p>
+          </div>
+          <div className={styles["about-box1"]}>
+            <svg
+              width="39"
+              height="27"
+              viewBox="0 0 39 27"
+              fill="none"
+              xmlns="http://www.w3.org/2000/svg"
+            >
+              <path
+                d="M27.5417 9.16667L32.8026 6.01009C34.4219 5.03848 35.2316 4.55268 35.8978 4.61479C36.4787 4.66894 37.0087 4.969 37.3539 5.43926C37.75 5.97856 37.75 6.92277 37.75 8.81121V17.6888C37.75 19.5771 37.75 20.5214 37.3539 21.0608C37.0087 21.531 36.4787 21.8311 35.8978 21.8852C35.2316 21.9473 34.4219 21.4616 32.8026 20.49L27.5417 17.3333M7.53333 25.5H21.0083C23.2952 25.5 24.4387 25.5 25.3122 25.0549C26.0804 24.6635 26.7052 24.0388 27.0966 23.2705C27.5417 22.3971 27.5417 21.2535 27.5417 18.9667V7.53333C27.5417 5.24646 27.5417 4.10301 27.0966 3.22954C26.7052 2.4612 26.0804 1.83653 25.3122 1.44506C24.4387 1 23.2952 1 21.0083 1H7.53333C5.24646 1 4.10301 1 3.22954 1.44506C2.4612 1.83653 1.83653 2.4612 1.44506 3.22954C1 4.10301 1 5.24644 1 7.53333V18.9667C1 21.2535 1 22.3971 1.44506 23.2705C1.83653 24.0388 2.4612 24.6635 3.22954 25.0549C4.10301 25.5 5.24644 25.5 7.53333 25.5Z"
+                stroke="white"
+                strokeWidth="2"
+                strokeLinecap="round"
+                strokeLinejoin="round"
+              />
+            </svg>
+            <p className={styles["about-team-name"]}>AMV Team</p>
+            <p>
+              Lorem ipsum odor amet, consectetuer adipiscing elit. Odio iaculis
+              primis nisi imperd
+            </p>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+}
