+++ components/icons/RightArrowIcon.jsx
@@ -1,0 +1,19 @@
+export default function RightArrowIcon() {
+  return (
+    <svg
+      width="14"
+      height="15"
+      viewBox="0 0 14 15"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M1 7.50006H13M13 7.50006L8 1.00006M13 7.50006L8 14.0001"
+        stroke="white"
+        strokeWidth="2"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      />
+    </svg>
+  );
+}
